HttpResponse Header info is now placed inside DTOs
Status code is in DTOs now along with the rest of the header information.
can call getHeaderProperties(); on a DTO to get Map<String, List<String>>
can call getStatusCode(); on a DTO to get the Integer status code
the responseCode is now an Integer

the Injector deserialize() method is now limited to only accept objects that are either SuperDTO or subclasses of SuperDTO, which also means that the deserialize() method will return SuperDTO (which can then be cast to the specific DTO you want).