DriverFactory class made
Please forgive the branch name, that is no longer the goal of this branch.
There is now a DriverFactory static method getDriver() that you should be able to call
that will take input from the webTesting.properties file to get you the driver you need.
This feature currently only supports chrome, firefox, and safari. You do not need a driver
executable on mac for the Safari browser.

Please when closing the driver returned from this method only use the quit() method.
The quit method will close the driver and any browser windows opened by the driver.
using both quit() and exit() will throw an exception and cause your tests to fail.

This feature has currently ONLY been tested on mac, so I still need a windows user to
verify that it will work correctly in that environment (should do hopefully? xD)

If there is any issue with creating the driver itself then getDriver() method will
return null and print an error message to the console.