Test cases 7 and 19
HomePage modified. Also it appears because my BrandProductsSideBarNav.feature has 2 Scenarios and 2 Givens that I essentially had to build up the given again rather than continuing from the previously defined properties which is fine and it works, just >;v

A bigger thing to note is that the only way I have found to successfully avoid issues with the ad that covers the whole browser window would be to run in headless mode. All tests pass in headless mode, but not otherwise which I mean it's not ideal, but it's what I got until we find another way to handle those pesky advertisements.